+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,15 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Admin on 19.05.2023.
+//
+
+import Foundation
+
+/// структура для массива со списком моковых вопросов
+struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
